requirements.txt
# Python Requirements for Autonomous Robot Project
# ================================================
# 
# For Jetson Nano:
#   pip3 install -r requirements.txt
#
# For Training PC:
#   pip install -r requirements.txt

# Core packages
numpy>=1.19.0
opencv-python>=4.5.0
pyserial>=3.5

# Deep Learning (PyTorch)
# Note: For Jetson Nano, install PyTorch from NVIDIA's wheels:
# https://forums.developer.nvidia.com/t/pytorch-for-jetson/
torch>=1.9.0
torchvision>=0.10.0

# Training only (not needed on Jetson)
pandas>=1.3.0
scikit-learn>=0.24.0
joblib>=1.0.0

# Optional: For better progress bars during training
# tqdm>=4.60.0
